requirements.txt
-i https://pypi.org/simple
appwrite==2.0.2
attrs==23.1.0; python_version >= '3.7'
automat==22.10.0
certifi==2023.7.22; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0; python_full_version >= '3.7.0'
constantly==15.1.0
cryptography==41.0.3; python_version >= '3.7'
cssselect==1.2.0; python_version >= '3.7'
filelock==3.12.2; python_version >= '3.7'
hyperlink==21.0.0
idna==3.4; python_version >= '3.5'
incremental==22.10.0
itemadapter==0.8.0; python_version >= '3.7'
itemloaders==1.1.0; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
lxml==4.9.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
packaging==23.1; python_version >= '3.7'
parsel==1.8.1; python_version >= '3.7'
protego==0.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydispatcher==2.0.7; platform_python_implementation == 'CPython'
pyopenssl==23.2.0; python_version >= '3.6'
python-dotenv==1.0.0
queuelib==1.6.2; python_version >= '3.5'
requests==2.31.0; python_version >= '3.7'
requests-file==1.5.1
scrapy==2.9.0
service-identity==23.1.0; python_version >= '3.8'
setuptools==68.0.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tldextract==3.4.4; python_version >= '3.7'
twisted==22.10.0; python_full_version >= '3.7.1'
typing-extensions==4.7.1; python_version >= '3.7'
urllib3==2.0.4; python_version >= '3.7'
w3lib==2.1.1; python_version >= '3.7'
zope.interface==6.0; python_version >= '3.7'
